Use long long int for npy_intp in reshape_funcs.cu
This is 64 bits on Linux and Windows, which corrects an issue with the
prototypes not matching on Windows.

+++ numba_cuda/numba/cuda/reshape_funcs.cu
@@ -4,7 +4,7 @@
  */
 #define NPY_MAXDIMS 32
 
-typedef long int npy_intp;
+typedef long long int npy_intp;
 
 extern "C" __device__ int
 nocopy_empty_reshape(npy_intp nd, const npy_intp *dims, const npy_intp *strides,
